Extract abstract class and modernize
